Updated build process to use subfolders & changed old makefile names to be compatible with new build Added support for macos, including build a 64bit version as Catalina has dropped support for 32bit Changed interface to ipipe-kill so that the default action can be configured to send a SIGTERM rather than a SIGKILL and if n is given this is the signal sent Fixed issue with signal handlers being installed to late when ME is run as a 'shell script' process Fixed an ipipe terminal size issue and allowed the width to be set via new %ipipe-width user variable Added support for X11 utf8 clipboard copy, which can lead to unmappable chars being converted to 0x1c
git-svn-id: file:///media/bill/diskAshur2/backup/svn/microemacs/trunk@2334 a259ec66-be10-11ea-b557-b34f8aa0b108

+++ microemacs/src/print.c
@@ -1631,7 +1631,7 @@ printRegion (int f, int n)
 
     if ((startLineNo=frameCur->windowCur->dotLineNo-frameCur->windowCur->markLineNo) == 0)
         return meTRUE;
-    numLines = abs (startLineNo);
+    numLines = abs((int) startLineNo);
     if (startLineNo < 0)
     {
         startLineNo = frameCur->windowCur->dotLineNo;
